Split mock-wallet into a reusable provider core
The package's only export, installMockWallet, bundled the EIP-1193
request handler, the Playwright injection, and the EIP-6963
announcement into a single Playwright-only function — unusable by
anything without a Page.

Extract the request handler (createMockWalletProvider) and the
announce helper (announceEip6963Provider) into host-agnostic modules
so an in-browser dev wallet can reuse the exact same wallet behavior,
keeping installMockWallet as a thin Playwright consumer with identical
behavior. The from-address guard now uses viem's isAddressEqual per
repo convention. The tsconfig gains allowImportingTsExtensions because
the package now has internal .ts imports (nodenext requires the flag).
